requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
colorzero==2.0
gpiozero==1.6.2
paho-mqtt==1.6.1
pigpio==1.78
rpi.gpio==0.7.1
setuptools==62.1.0; python_version >= '3.7'
